Refactor to local_plannng_benchmark
